Refractor: update package paths for better organization delete 'error' package, move to 'excpetion' package move 'RedisConfig' class to 'config' package
