+++ src/main/java/umc/duckmelang/domain/auth/security/jwt/JwtTokenProvider.java
@@ -10,7 +10,7 @@ import org.springframework.security.config.annotation.web.configuration.EnableWe
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Component;
 import umc.duckmelang.global.apipayload.code.status.ErrorStatus;
-import umc.duckmelang.global.error.exception.TokenException;
+import umc.duckmelang.global.apipayload.exception.TokenException;
 
 import java.security.Key;
 import java.util.Collections;
