+++ src/main/java/umc/duckmelang/global/apipayload/exception/AuthException.java
@@ -1,8 +1,8 @@
-package umc.duckmelang.global.error.exception;
+package umc.duckmelang.global.apipayload.exception;
 
 
 import umc.duckmelang.global.apipayload.code.BaseErrorCode;
-import umc.duckmelang.global.error.GeneralException;
+import umc.duckmelang.global.apipayload.GeneralException;
 
 public class AuthException extends GeneralException {
     public AuthException(BaseErrorCode errorCode){ super(errorCode); }
