+++ src/main/java/umc/duckmelang/global/apipayload/exception/MemberException.java
@@ -1,7 +1,7 @@
-package umc.duckmelang.global.error.exception;
+package umc.duckmelang.global.apipayload.exception;
 
 import umc.duckmelang.global.apipayload.code.BaseErrorCode;
-import umc.duckmelang.global.error.GeneralException;
+import umc.duckmelang.global.apipayload.GeneralException;
 
 public class MemberException extends GeneralException {
     public MemberException(BaseErrorCode errorCode) {super(errorCode);}
