+++ src/main/java/umc/duckmelang/global/config/RedisConfig.java
@@ -1,4 +1,4 @@
-package umc.duckmelang.global.infra;
+package umc.duckmelang.global.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
