+++ src/main/java/umc/duckmelang/global/temp/service/TempQueryServiceImpl.java
@@ -3,7 +3,7 @@ package umc.duckmelang.global.temp.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import umc.duckmelang.global.apipayload.code.status.ErrorStatus;
-import umc.duckmelang.global.error.handler.TempHandler;
+import umc.duckmelang.global.apipayload.exception.handler.TempHandler;
 
 @Service
 @RequiredArgsConstructor
